updated rbc center offset to only apply when atleast one of the existing bars on the chart are visible
we were getting terrible x-axis label scaling on the multichart when used with bars and lines while the max width of the chart is very small and only the lines on the chart are enabled(all bars disabled/hidden). I have found that this was caused by the range band center offset is being applied for the invisible bars when it is unecessary and it screws with the scaling/translation of the interactive/line layer. For info the range band center offset is used to make it so the point on the lines are inline with the center of the group of bars(not at the start of each group like it is by default).